Refactor styled form components into own file
