+++ packages/frontend/src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { ThemeProvider } from 'styled-components';
 import { GlobalStyle, theme } from './ui/theme';
-import UserLogin from './components/UserLogin';
+import UserLogin from './components/UserSignup';
 
 function App() {
   return (
