+++ packages/frontend/src/ui/components/StyledFormComponents.js
@@ -1,0 +1,31 @@
+
+import styled from 'styled-components';
+
+
+export const SubmitButton = styled.button`
+  align-self: center;
+`;
+
+export const FormGroup = styled.div`
+  padding: 0.5rem;
+  justify-content: space-between;
+  display: flex;
+  flex-direction: column;
+`;
+
+export const SignupFormContainer = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+  width: 100%;
+`;
+
+export const InputLabel = styled.label`
+  padding-right: 0.5rem;
+`;
+
+export const TextInput = styled.input`
+  padding: 0.5rem;
+  width: 100%;
+`;
